+++ modules/terraform-aci-imported-l4l7-device/outputs.tf
@@ -1,0 +1,4 @@
+output "dn" {
+  value       = aci_rest_managed.vnsLDevIf.id
+  description = "Distinguished name of `vnsLDevIf` object."
+}